Start splitting unit and API regression tests
